# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FC Rating API",
    "version": "0.1.0",
    "description": "Elo-based match rating API for a friend group's EA Sports FC league."
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fc_session"
      }
    },
...
  },
  "paths": {
...
    "/players": {
      "post": {
        "operationId": "createPlayer",
        "summary": "Create a player (requires admin role)",
        "tags": [
          "players"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "avatarUrl": {
                    "type": "string",
                    "format": "uri"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "avatarUrl": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "isActive": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "avatarUrl",
                    "isActive"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listPlayers",
        "summary": "List players",
        "tags": [
          "players"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "in": "query",
            "name": "active",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "avatarUrl": {
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "isActive": {
                        "type": "boolean"
                      },
                      "lastPlayedAt": {
                        "anyOf": [
                          {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "avatarUrl",
                      "isActive",
                      "lastPlayedAt"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/players/{id}": {
      "get": {
        "operationId": "getPlayer",
        "summary": "Get a player profile",
        "tags": [
          "players"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "playerId": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    "rating": {
                      "type": "number"
                    },
                    "gamesPlayed": {
                      "type": "number"
                    },
                    "wins": {
                      "type": "number"
                    },
                    "draws": {
                      "type": "number"
                    },
                    "losses": {
                      "type": "number"
                    },
                    "form": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "W",
                          "L",
                          "D"
                        ]
                      }
                    },
                    "streak": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "result": {
                              "type": "string",
                              "enum": [
                                "W",
                                "L",
                                "D"
                              ]
                            },
                            "length": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "result",
                            "length"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "bestStreak": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "result": {
                              "type": "string",
                              "enum": [
                                "W",
                                "L",
                                "D"
                              ]
                            },
                            "length": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "result",
                            "length"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "goalsFor": {
                      "type": "number"
                    },
                    "goalsAgainst": {
                      "type": "number"
                    },
                    "isProvisional": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "playerId",
                    "name",
                    "isActive",
                    "createdAt",
                    "rating",
                    "gamesPlayed",
                    "wins",
                    "draws",
                    "losses",
                    "form",
                    "streak",
                    "bestStreak",
                    "goalsFor",
                    "goalsAgainst",
                    "isProvisional"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updatePlayer",
        "summary": "Update a player (requires admin role)",
        "tags": [
          "players"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "avatarUrl": {
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uri"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "avatarUrl": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "isActive": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "avatarUrl",
                    "isActive"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deletePlayer",
        "summary": "Delete a player who has never played a match (requires admin role)",
        "tags": [
          "players"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/players/{id}/rating-history": {
      "get": {
        "operationId": "getPlayerRatingHistory",
        "summary": "Get a player's rating history",
        "tags": [
          "players"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "matchId": {
                        "type": "string"
                      },
                      "sequence": {
                        "type": "number"
                      },
                      "playedAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                      },
                      "before": {
                        "type": "number"
                      },
                      "after": {
                        "type": "number"
                      },
                      "delta": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "matchId",
                      "sequence",
                      "playedAt",
                      "before",
                      "after",
                      "delta"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/players/{id}/matches": {
      "get": {
        "operationId": "listPlayerMatches",
        "summary": "List a player's matches, cursor-paginated",
        "tags": [
          "players"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "sequence": {
                            "type": "number"
                          },
                          "homePlayerId": {
                            "type": "string"
                          },
                          "awayPlayerId": {
                            "type": "string"
                          },
                          "homeScore": {
                            "type": "number"
                          },
                          "awayScore": {
                            "type": "number"
                          },
                          "isVoid": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "id",
                          "sequence",
                          "homePlayerId",
                          "awayPlayerId",
                          "homeScore",
                          "awayScore",
                          "isVoid"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "number",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [],
...
}